user and memes routes are protected

+++ app/api/memes/protected/route.ts
@@ -5,41 +5,90 @@ import User from "@/lib/models/users";
 import connect from "@/lib/db";  
 import { Types } from "mongoose";
 
+
+
+
+
+export const GET = async (request : Request) =>{
+
+
+try{
+
+const {searchParams} = new URL(request.url);
+const userId = searchParams.get("user")
+
+
+if(!userId || userId === "") {
+  return new NextResponse("userid missing" , {status:404});
+}
+
+if(!mongoose.Types.ObjectId.isValid(userId)){
+  return new NextResponse("User Id isnt valid" , {status:404});
+}
+
+ const user = await User.findById(userId);
+
+ if(!user){
+   return new NextResponse("User not found" , {status:404});
+ }
+
+
+
+await connect();
+
+
+const memes = await Meme.find();
+return new NextResponse(JSON.stringify(memes) , {status:200})
+
+
+}catch(e:any){
+
+    return new NextResponse("Error in fetching memes" +e.message , {status:500});
+
+}
+
+
+
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 export const POST = async (request : Request)=>{
 
 try{
 
 const {searchParams} = new URL(request.url);
-const userId = searchParams.get('userId');
+const loggeduserid = request.headers.get("loggeduserid")
 
-if(!userId){
-
-  return new NextResponse('User id not found' , {status:400})
-
-}
-
-if(!mongoose.Types.ObjectId.isValid(userId)){
-
-   return new NextResponse('User id is not valid' , {status:400})
-}
 
 
 await connect();
 
-const user = await User.findById(userId);
-
-if(!user){
-
- return new NextResponse('User  not found' , {status:400})
-}
 
 
 const body = await request.json();
+const {userid} = body;
 
 if(!body){
  return new NextResponse('Body not found' , {status:400})
 }
 
+if(userid !== loggeduserid){
+ return new NextResponse('why the fuck u are trying to access another users ' , {status:404})
+}
 
 const newMeme = new Meme(body);
 
@@ -68,31 +117,26 @@ export const PATCH = async (request : Request ) =>{
 try{
 
 const {searchParams} = new URL(request.url);
-const userId = searchParams.get('userId');
 const memeId = searchParams.get('memeId');
+const loggeduserid = request.headers.get("loggeduserid")
 
-if (!userId || !memeId) {
+if ( !memeId) {
       return new NextResponse(
-        JSON.stringify({ message: "missing userId " }),
+        JSON.stringify({ message: "missing memeId " }),
         { status: 400 }
       );
     }
 
 
-if(!Types.ObjectId.isValid(userId) || !Types.ObjectId.isValid(memeId)){
+if( !Types.ObjectId.isValid(memeId)){
 
-    return new NextResponse(JSON.stringify({message : 'Invalide  id' } ) ,{status : 400})
+    return new NextResponse(JSON.stringify({message : 'Invalide  memeid' } ) ,{status : 400})
 }
 
 
 await connect();
 
-const user = await User.findById(userId);
 
-if(!user){
-
- return new NextResponse('User  not found' , {status:400})
-}
 
 const meme = await Meme.findById(memeId);
 
@@ -100,6 +144,12 @@ if(!meme){
 
  return new NextResponse('Meme  not found' , {status:400})
 }
+
+
+if(meme.userid.toString() !== loggeduserid){
+ return new NextResponse('why the fuck u are trying to access another users ' , {status:404})
+}
+
 
 const body = await request.json();
 
@@ -148,6 +198,8 @@ try {
 
 const {searchParams} = new URL(request.url);
 const memeId = searchParams.get("memeId");
+const loggeduserid = request.headers.get("loggeduserid")
+
 
 if (!memeId) {
       return new NextResponse(
@@ -163,6 +215,18 @@ if(!Types.ObjectId.isValid(memeId)){
 }
 
 await connect();
+
+const meme = await Meme.findById(memeId);
+
+if(!meme){
+  return new NextResponse(JSON.stringify({ message: "Meme not found" }), {
+        status: 404,
+      });
+}
+
+if(meme.userid.toString() !== loggeduserid){
+ return new NextResponse('why the fuck u are trying to access another users ' , {status:404})
+}
 
 const deletedmeme = await Meme.findByIdAndDelete(memeId);
